Support acpi event files for the GPIO reboot interrupt
For both BF2 and BF3, there is a GPIO controlled by the BMC,
responsible to trigger the DPU graceful reboot.

Until now the requirement was a soft reset (forced reboot) but this
requirement has been changed following the request from a customer.
So for all distros and both BF2/BF3, change the behavior to graceful
reboot. To achieve that, create acpi event files:
/etc/acpi/events/mlnx-powerconf and
/etc/acpi/actions/rebootcontrol.

This change works hand in hand with the pwr-mlxbf driver.

RM #3591305